+++ Cargo.toml
@@ -39,7 +39,7 @@ tokio-io = "0.1"
 [target.'cfg(not(target_os = "emscripten"))'.dependencies]
 libp2p-dns = { path = "./transports/dns" }
 libp2p-mdns = { path = "./misc/mdns" }
-libp2p-tcp-transport = { path = "./transports/tcp" }
+libp2p-tcp = { path = "./transports/tcp" }
 
 [target.'cfg(target_os = "emscripten")'.dependencies]
 stdweb = { version = "0.1.3", default-features = false }
