+++ protocols/ping/Cargo.toml
@@ -24,6 +24,6 @@ tokio-timer = "0.2.6"
 void = "1.0"
 
 [dev-dependencies]
-libp2p-tcp-transport = { path = "../../transports/tcp" }
+libp2p-tcp = { path = "../../transports/tcp" }
 tokio = "0.1"
 tokio-tcp = "0.1"
